Removed GeneratedTEKImpl class; Former interface GeneratedTEK is now a class.
