+++ coralibre-sdk/sdk/src/main/java/org/coralibre/android/sdk/internal/database/ppcp/model/GeneratedTEK.java
@@ -1,9 +1,29 @@
 package org.coralibre.android.sdk.internal.database.ppcp.model;
 
+
 import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
 
-public interface GeneratedTEK {
-    ENNumber getInterval();
 
-    byte[] getKey();
+public class GeneratedTEK {
+
+
+    private final ENNumber interval;
+
+    private final byte[] key;
+
+
+
+    public GeneratedTEK(ENNumber interval, byte[] key) {
+        this.interval =interval;
+        this.key = key;
+    }
+
+
+    public ENNumber getInterval() {
+        return interval;
+    }
+
+    public byte[] getKey() {
+        return key;
+    }
 }
